+++ src/environment.rs
@@ -1,89 +1,103 @@
+use crate::ast::match_::Match;
 use std::cell::RefCell;
 use std::fmt::Debug;
 use std::rc::Rc;
-use crate::ast::match_::Match;
 
 #[derive(Debug, PartialEq, Clone)]
 pub enum Env<V>
-where V: EnvVal {
+where
+    V: EnvVal,
+{
     Empty,
     Entry {
         ident: String,
         value: V,
-        parent: Rc<RefCell<Env<V>>>
-    }
+        parent: Rc<RefCell<Env<V>>>,
+    },
 }
 
 pub type EnvWrapper<E> = Rc<RefCell<E>>;
 
-impl<V> Env<V> 
-where V: EnvVal {
+impl<V> Env<V>
+where
+    V: EnvVal,
+{
     pub fn empty() -> EnvWrapper<Self> {
         Rc::new(RefCell::new(Env::Empty))
     }
-    
-    pub fn associate(pattern: Match, value: V, parent: &EnvWrapper<Self>) -> Result<EnvWrapper<Self>, String> {
+
+    pub fn associate(
+        pattern: Match,
+        value: V,
+        parent: &EnvWrapper<Self>,
+    ) -> Result<EnvWrapper<Self>, String> {
         let parent = Rc::clone(parent);
-        value.unwrap_matches(&pattern).map(
-            |matches|
-            matches.into_iter().fold(
-                parent,
-                |parent, (ident, value)|
-                    Env::associate_ident(ident, value, parent),
-            )
-        )
+        value.unwrap_matches(&pattern).map(|matches| {
+            matches.into_iter().fold(parent, |parent, (ident, value)| {
+                Env::associate_ident(ident, value, parent)
+            })
+        })
     }
-    
     pub fn associate_ident(ident: String, value: V, parent: EnvWrapper<Self>) -> EnvWrapper<Self> {
-        Rc::new(RefCell::new(Env::Entry { ident, value, parent }))
+        Rc::new(RefCell::new(Env::Entry {
+            ident,
+            value,
+            parent,
+        }))
     }
-    
     pub fn get(env: &EnvWrapper<Self>, ident: &str) -> Option<V> {
         match *env.borrow() {
             Env::Empty => None,
-            Env::Entry { ident: ref id, ref value, ref parent } =>
+            Env::Entry {
+                ident: ref id,
+                ref value,
+                ref parent,
+            } => {
                 if ident == id {
                     Some(value.clone())
                 } else {
                     Env::get(parent, ident)
                 }
+            }
         }
     }
-    
     pub fn get_evaluated_value(env: &EnvWrapper<Self>) -> Result<V, String> {
         match *env.borrow() {
             Env::Empty => Err("No variables are declared".to_string()),
-            Env::Entry { ref value, ..} if value.is_evaluated()
-                => Ok(value.clone()),
-            Env::Entry { .. } => Err("Variable is not evaluated".to_string())
+            Env::Entry { ref value, .. } if value.is_evaluated() => Ok(value.clone()),
+            Env::Entry { .. } => Err("Variable is not evaluated".to_string()),
         }
     }
-    
     pub fn with_parent(env: &EnvWrapper<Self>, env_parent: &EnvWrapper<Self>) -> EnvWrapper<Self> {
         match *env.borrow() {
             Env::Empty => Rc::clone(env_parent),
-            Env::Entry { ref ident, ref value, ref parent} => 
-                Rc::new(RefCell::new(Env::Entry {
-                    ident: ident.to_string(),
-                    value: value.clone(),
-                    parent: Env::with_parent(parent, env_parent)
-                }))
+            Env::Entry {
+                ref ident,
+                ref value,
+                ref parent,
+            } => Rc::new(RefCell::new(Env::Entry {
+                ident: ident.to_string(),
+                value: value.clone(),
+                parent: Env::with_parent(parent, env_parent),
+            })),
         }
     }
-    
-    // Only should be used with declarations to solve recursive problem
+    // Only should be used with declarations and delayed values to solve recursive problem
     // Also used with evaluation of lazy values
     // I strongly dislike the fact that I have to do this
     pub fn set_value(env: &EnvWrapper<Self>, value: V) {
         let mut inner_env = env.borrow_mut();
         let new_inner_env = match *inner_env {
             Env::Empty => Env::Empty,
-            Env::Entry {ref ident, ref parent, ..} =>
-                Env::Entry {
-                    ident: ident.to_string(),
-                    value,
-                    parent: Rc::clone(parent)
-                }
+            Env::Entry {
+                ref ident,
+                ref parent,
+                ..
+            } => Env::Entry {
+                ident: ident.to_string(),
+                value,
+                parent: Rc::clone(parent),
+            },
         };
         *inner_env = new_inner_env;
     }
@@ -97,7 +111,6 @@ pub trait EnvVal: Clone + Debug {
 #[cfg(test)]
 mod tests {
     use super::*;
-    
     basic_test! {
         no_parent_test
         Env::get(&Env::associate(
@@ -111,7 +124,6 @@ mod tests {
             &Env::empty()
         ).unwrap(), "b") => None
     }
-    
     basic_test! {
         with_parent_test
         Env::get(&Env::associate(Match::ident("a"), DummyValue::int("a", 1),
@@ -124,7 +136,6 @@ mod tests {
             &Env::associate(Match::ident("b"), DummyValue::int("b", 2), &Env::empty()).unwrap())
             .unwrap(), "c") => None
     }
-    
     basic_test! {
         with_tuple_test
         Env::get(&Env::associate(
@@ -150,21 +161,14 @@ mod tests {
             &Env::empty()
         ).unwrap(), "b") => Some(DummyValue::Int(2))
     }
-    
     #[test]
     fn setter_test() {
         let expected = Some(DummyValue::Int(2));
-        let env = Env::associate_ident(
-                "a".to_string(),
-                DummyValue::Int(1),
-                Env::empty()
-        );
+        let env = Env::associate_ident("a".to_string(), DummyValue::Int(1), Env::empty());
         Env::set_value(&env, DummyValue::Int(2));
         let actual = Env::get(&env, "a");
-        
         assert_eq!(expected, actual);
     }
-    
     basic_test! {
         get_value_test
         Env::<DummyValue>::get_evaluated_value(&Env::empty())
@@ -180,7 +184,6 @@ mod tests {
             Env::empty()
         )) => Err("Variable is not evaluated".to_string())
     }
-    
     // Basically checking the (Haskell) monoidal properties
     // of `Env`, with `EnvWrapper<Env>` as the monoid,
     // `Env::with_parent` as the `++` operator,
@@ -234,43 +237,37 @@ mod tests {
             )
         )
     }
-    
     #[derive(Debug, Clone, PartialEq)]
     enum DummyValue {
         Int(u32),
         Container(Vec<(String, Self)>),
-        Delayed(u32)
+        Delayed(u32),
     }
-    
     impl DummyValue {
         fn new(contains: Vec<(&str, u32)>) -> Self {
             DummyValue::Container(
-                contains.into_iter()
-                .map(
-                    |(ident, val)|
-                    (ident.to_string(), DummyValue::Int(val))
-                ).collect()
+                contains
+                    .into_iter()
+                    .map(|(ident, val)| (ident.to_string(), DummyValue::Int(val)))
+                    .collect(),
             )
         }
-        
         fn int(ident: &str, val: u32) -> Self {
             DummyValue::Container(vec![(ident.to_string(), DummyValue::Int(val))])
         }
     }
-    
     impl EnvVal for DummyValue {
         fn unwrap_matches(&self, _pattern: &Match) -> Result<Vec<(String, Self)>, String> {
             match self {
                 DummyValue::Container(a) => Ok(a.clone()),
                 DummyValue::Int(_) => Err("Int value can't be unwrapped!".to_string()),
-                DummyValue::Delayed(_) => Err("Int value can't be unwrapped!".to_string())
+                DummyValue::Delayed(_) => Err("Int value can't be unwrapped!".to_string()),
             }
         }
-        
         fn is_evaluated(&self) -> bool {
             match self {
                 DummyValue::Delayed(_) => false,
-                _ => true
+                _ => true,
             }
         }
     }