+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* Delay expressions
+  * Delay the evaluation of an expression until later
+  * `delay <ident> = <expr> in <expr>`
+  * Only accepts identifiers, no other match expressions
+  * Expression is evaluated when the variable is used
+
 ### Changed
 * Declarations are not evaluated until needed
   * Any declarations referencing future declarations now work